requirements.txt
docopt
sphinx_py3doc_enhanced_theme
coverage
irc-parser-tests
